requirements.txt
# Core
fastapi
uvicorn[standard]
pydantic
pydantic-settings

# HTTP Clients
httpx
openai

# Redis
redis
hiredis

# Configuration
pyyaml
python-dotenv

# Logging
structlog

# Testing
pytest
pytest-asyncio
pytest-mock

# Development
black
flake8